# Short circuit volume checker if the pod is not requesting any volumes