**What type of PR is this?**
/kind cleanup

This is an optimization.

**What this PR does / why we need it**:
Volume checker predicate function turns out to take about 5% of the scheduler execution time in our benchmarks while pods in those tests do not request any volumes. This PR short circuits the logic for pods that do not need any volumes.

Before:
```
BenchmarkScheduling/5000Nodes/1000Pods-12         	   10000	   9140564 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	114.688s
```
After:
```
pkg: k8s.io/kubernetes/test/integration/scheduler_perf
BenchmarkScheduling/5000Nodes/1000Pods-12         	   10000	   8541605 ns/op
PASS
ok  	k8s.io/kubernetes/test/integration/scheduler_perf	108.582s
```

**Which issue(s) this PR fixes**:
<!-- 
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--  
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
NONE
```

/sig scheduling
/assign @msau42 @cofyc