@cofyc 2019-02-02
LGTM
It makes sense to skip `FindPodVolumes` in predicate function when it knows there is no need to find pod volumes for it. Prometheus metrics in volume binding will record latency only for pods which needs volume scheduling, I think this is acceptable.
btw, `TestSchedulerWithVolumeBinding` fails, I guess we need to pass a pod with PVCs to test VolumeBinder now