@bsalamat 2019-02-02
Thanks, @cofyc!
I added a PVC to the pod and the client to make the test happy.